macos_requirements.txt

absl-py                 1.4.0
ase                     3.22.1
astunparse              1.6.3
cachetools              5.3.0
certifi                 2022.12.7
charset-normalizer      3.0.1
cloudpickle             2.2.1
contourpy               1.0.7
cycler                  0.11.0
decorator               5.1.1
flatbuffers             1.12
fonttools               4.38.0
gast                    0.4.0
google-auth             2.16.1
google-auth-oauthlib    0.4.6
google-pasta            0.2.0
googledrivedownloader   0.4
grpcio                  1.51.3
h5py                    3.8.0
idna                    3.4
importlib-metadata      6.0.0
importlib-resources     5.12.0
isodate                 0.6.1
Jinja2                  3.1.2
joblib                  1.2.0
keras                   2.9.0
Keras-Preprocessing     1.1.2
kiwisolver              1.4.4
libclang                15.0.6.1
llvmlite                0.39.1
Markdown                3.4.1
MarkupSafe              2.1.2
matplotlib              3.7.0
networkx                2.6.3
numba                   0.56.4
numpy                   1.23.5
oauthlib                3.2.2
opt-einsum              3.3.0
packaging               23.0
pandas                  1.5.3
Pillow                  9.4.0
pip                     23.0.1
protobuf                3.19.6
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pynndescent             0.5.8
pyparsing               3.0.9
python-dateutil         2.8.2
python-louvain          0.16
pytz                    2022.7.1
rdflib                  6.2.0
requests                2.28.2
requests-oauthlib       1.3.1
rsa                     4.9
scikit-learn            1.2.1
scipy                   1.10.1
seaborn                 0.12.2
setuptools              65.6.3
six                     1.16.0
tensorboard             2.9.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.1
tensorflow-estimator    2.9.0
tensorflow-macos        2.9.2
tensorflow-metal        0.7.1
tensorflow-probability  0.7.0
termcolor               1.1.0
threadpoolctl           3.1.0
torch                   1.9.0
torch-geometric         1.7.0
torchvision             0.10.0
tqdm                    4.59.0
typing_extensions       4.5.0
umap-learn              0.5.1
urllib3                 1.26.14
Werkzeug                2.0.3
wheel                   0.38.4
wrapt                   1.14.1
zipp                    3.6.0